paycart

+++ cart_2.php
@@ -18,10 +18,13 @@ $row = $result->fetch(PDO::FETCH_ASSOC);
 </head>
 <body>
     <div class="container">
-        <form action="" id="form1" method="post">
+        <form action="print_product.php" id="form1" method="post">
         <div class="row">
             <div class="col-10">
                 <table class="table table-hover"> 
+                <div class="alert alert-primary h4 text-center mt-3" role="alert">
+                ตระกร้าสินค้า
+            </div>
                     <tr>
                         <th>ลำดับที่</th>
                         <th>ชื่อสินค้า</th>
@@ -53,7 +56,7 @@ $row = $result->fetch(PDO::FETCH_ASSOC);
                                 <td><?= $price ?></td>  
                                 <td><?= $qty ?></td>  
                                 <td><?= $subtotal ?></td> 
-                                <td><a href="product_delete.php?Line=<?= $i ?>"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+                                <td><a href="product_delete.php?Line=<?= $i ?>" onclick="return confirm('คุณต้องการลบสินค้าในตะกร้าหรือไม่?')"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16">
   <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5m-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5M4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06m6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528M8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5"/>
 </svg></a></td><!-- แสดงราคารวม -->
                             </tr>
@@ -67,14 +70,31 @@ $row = $result->fetch(PDO::FETCH_ASSOC);
                         <td class="text-center"><?= $sumprice ?></td>
                         <td>บาท</td>
                     </tr>
+                    <?php if($_SESSION["intLine"] <= 0): ?>
+                    <tr>
+                        <td colspan="6" class="text-center">ไม่มีสินค้าในตะกร้า</td>
+                    </tr>
+                    <?php endif; ?>
                 </table>
                 <div style="text-align: right">
                 <a href="productCatalogPagination.php" type="button" class="btn btn-primary">เลือกสินค้า</a> |
-                <button type="button" class="btn btn-danger">ยืนยันการสั่งซื้อ</button>
+                <button type="button" class="btn btn-danger" onclick="confirmCheckout()">ยืนยันการสั่งซื้อ</button>
                 </div>
             </div>
         </div>
         </form> 
     </div>
+
+<script>
+function confirmCheckout() {
+    if(<?php echo $_SESSION["intLine"]; ?> <= 0) {
+        alert("กรุณาเลือกสินค้าก่อนทำการสั่งซื้อ");
+    } else {
+        // ทำการส่งข้อมูลไปยังหน้าที่ต้องการทำการสั่งซื้อ
+        document.getElementById("form1").submit();
+    }
+}
+</script>
+
 </body>
 </html>
